+++ src/main/java/com/panos/interfaces/ControllerListener.java
@@ -2,7 +2,7 @@ package com.panos.interfaces;
 
 // Interface for Controllers, in case we use different controllers
 public interface ControllerListener {
-    void onAxisChange(float lx, float ly, float rx, float ry);
+    void onAxisChange(double lx, double ly, double rx, double ry);
 
     void onButtonPress(int button, boolean isPressed);
 }
